extract read-elm-config.go from read-config.go
