+++ src/read-config.go
@@ -80,55 +80,6 @@ func decodeConfig(root interface{}) (config, error) {
 	return config, nil
 }
 
-func decodeElmConfig(root interface{}) (elmConfig, error) {
-	var ret elmConfig
-
-	dependencies, err := getObjectProperty(root, "elm.json", "dependencies")
-	if err != nil {
-		return ret, err
-	}
-
-	direct, err := getObjectProperty(dependencies, "elm.json['dependencies']", "direct")
-	if err != nil {
-		return ret, err
-	}
-
-	directObject, ok := direct.(map[string]interface{})
-	if !ok {
-		return ret, errors.New("elm.json['dependencies']['direct'] is not an object")
-	}
-
-	elmCoreVersion, ok := directObject["elm/core"]
-	if !ok {
-		return ret, errors.New("elm.json['dependencies']['direct']['elm/core'] not found")
-	}
-
-	elmCoreVersionString, ok := elmCoreVersion.(string)
-	if !ok {
-		return ret, errors.New("elm.json['dependencies']['direct']['elm/core'] is not a string")
-	}
-
-	ret.elmCoreVersion, err = parseVersion(elmCoreVersionString)
-	if err != nil {
-		return ret, err
-	}
-
-	dictExtraVersion, ok := directObject["elm-community/dict-extra"]
-	var dictExtraVersionString string
-	if ok {
-		dictExtraVersionString, ok = dictExtraVersion.(string)
-	}
-	if ok {
-		dictExtraVersion, err := parseVersion(dictExtraVersionString)
-		if err != nil {
-			return ret, err
-		}
-		ret.dictExtraVersion = &dictExtraVersion
-	}
-
-	return ret, nil
-}
-
 func decodeConfigFromBlob(blob []byte) (config, error) {
 	var root interface{}
 	err := json.Unmarshal(blob, &root)
@@ -139,16 +90,6 @@ func decodeConfigFromBlob(blob []byte) (config, error) {
 	return decodeConfig(root)
 }
 
-func decodeElmConfigFromBlob(blob []byte) (elmConfig, error) {
-	var root interface{}
-	err := json.Unmarshal(blob, &root)
-	if err != nil {
-		return elmConfig{}, err
-	}
-
-	return decodeElmConfig(root)
-}
-
 func readConfig() (config, error) {
 	configJson, err := os.ReadFile("gen-elm-wrappers.json")
 	if err != nil {
@@ -156,13 +97,4 @@ func readConfig() (config, error) {
 	}
 
 	return decodeConfigFromBlob(configJson)
-}
-
-func readElmConfig() (elmConfig, error) {
-	elmJson, err := os.ReadFile("elm.json")
-	if err != nil {
-		return elmConfig{}, err
-	}
-
-	return decodeElmConfigFromBlob(elmJson)
 }
